generateUniqueSku를 유틸함수로 분리
